+++ lib/src/web_tokens/mono_black/color/surface/surface.dart
@@ -1,0 +1,158 @@
+/*
+ * SPDX-FileCopyrightText: Copyright 2025 LG Electronics Inc.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import 'package:flutter/material.dart' show Color;
+
+import '../../../../core_tokens/color_primitive.dart';
+import '../../../base/color/surface/surface_base.dart';
+
+class Surface extends SurfaceBase {
+  const Surface();
+
+  @override
+  Color get defaultHandle => ColorPrimitive.instance.mistGray60;
+  @override
+  Color get defaultIndicator => ColorPrimitive.instance.mistGray90;
+  @override
+  Color get defaultNotification => ColorPrimitive.instance.redOrange60;
+  @override
+  Color get defaultPlaceholder => ColorPrimitive.instance.mistGray60;
+  @override
+  Color get defaultTrack => ColorPrimitive.instance.mistGray40;
+  @override
+  Color get main => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get mainHover => ColorPrimitive.instance.mistGray95;
+  @override
+  Color get mainPressed => ColorPrimitive.instance.mistGray90;
+  @override
+  Color get sub => ColorPrimitive.instance.mistGray40;
+  @override
+  Color get accent => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get buttonPrimary => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get buttonPrimaryHover => ColorPrimitive.instance.mistGray90;
+  @override
+  Color get buttonPrimaryPressed => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get buttonSecondary => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get buttonSecondaryHover => ColorPrimitive.instance.mistGray90;
+  @override
+  Color get buttonSecondaryPressed => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get buttonTertiary => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get buttonTertiaryHover => ColorPrimitive.instance.mistGray90;
+  @override
+  Color get buttonTertiaryPressed => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get buttonIcon => ColorPrimitive.instance.mistGray40;
+  @override
+  Color get buttonIconPressed => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get chip => ColorPrimitive.instance.warmGray25;
+  @override
+  Color get chipSelected => ColorPrimitive.instance.mistGray90;
+  @override
+  Color get notificationCard => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get selectionActive => ColorPrimitive.instance.green60;
+  @override
+  Color get selectionInactive => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get textFieldDisabled => ColorPrimitive.instance.mistGray40;
+  @override
+  Color get toast => ColorPrimitive.instance.mistGray60;
+  @override
+  Color get white => ColorPrimitive.instance.white;
+  @override
+  Color get black => ColorPrimitive.instance.mistGray10;
+  @override
+  Color get defaultError => ColorPrimitive.instance.activeRed60;
+  @override
+  Color get defaultSuccess => ColorPrimitive.instance.green70;
+  @override
+  Color get defaultWarning => ColorPrimitive.instance.yellow80;
+  @override
+  Color get mainDisabled => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get badgeGradientStart => ColorPrimitive.instance.deepOrange45;
+  @override
+  Color get badgeGradientStop => ColorPrimitive.instance.activeRed50;
+  @override
+  Color get badgeGradientEnd => ColorPrimitive.instance.heritageRed35;
+  @override
+  Color get badgeRed => ColorPrimitive.instance.activeRed50;
+  @override
+  Color get badgeLightRed => ColorPrimitive.instance.activeRed95;
+  @override
+  Color get badgeOrange => ColorPrimitive.instance.orange50;
+  @override
+  Color get badgeLightOrange => ColorPrimitive.instance.orange95;
+  @override
+  Color get badgeGreen => ColorPrimitive.instance.green50;
+  @override
+  Color get badgeLightGreen => ColorPrimitive.instance.green95;
+  @override
+  Color get badgeMagenta => ColorPrimitive.instance.heritageRed45;
+  @override
+  Color get badgeLightMagenta => ColorPrimitive.instance.heritageRed95;
+  @override
+  Color get badgeLightGray => ColorPrimitive.instance.coolGray90;
+  @override
+  Color get badgeBlack => ColorPrimitive.instance.mistGray40;
+  @override
+  Color get buttonPrimaryDisabled => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get buttonLight => ColorPrimitive.instance.coolGray90;
+  @override
+  Color get buttonDark => ColorPrimitive.instance.mistGray60;
+  @override
+  Color get cardGray => ColorPrimitive.instance.mistGray35;
+  @override
+  Color get chipPink => ColorPrimitive.instance.activeRed30;
+  @override
+  Color get dialogPopup => ColorPrimitive.instance.mistGray20;
+  @override
+  Color get selectionDisabled => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get selectionActiveDisabled => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get selectionSwitchInactive => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get selectionOptionDisabled => ColorPrimitive.instance.mistGray20;
+  @override
+  Color get tab => ColorPrimitive.instance.mistGray45;
+  @override
+  Color get tabActive => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get badgeDarkGray => ColorPrimitive.instance.coolGray35;
+  @override
+  Color get footer => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get headerDark => ColorPrimitive.instance.black;
+  @override
+  Color get headerLight => ColorPrimitive.instance.black;
+  @override
+  Color get subDisabled => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get cardOrange => ColorPrimitive.instance.orange40;
+  @override
+  Color get selectionInactiveDisabled => ColorPrimitive.instance.mistGray50;
+  @override
+  Color get banner => ColorPrimitive.instance.neutralGray5;
+  @override
+  Color get card => ColorPrimitive.instance.black;
+  @override
+  Color get filter => ColorPrimitive.instance.mistGray20;
+  @override
+  Color get selectionOption => ColorPrimitive.instance.mistGray30;
+  @override
+  Color get tabSegmented => ColorPrimitive.instance.black;
+  @override
+  Color get subSupporting => ColorPrimitive.instance.mistGray45;
+}