+++ packages/webos-tokens/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The following is a curated list of changes in the webos tokens module, newest changes on the top.
 
+## [0.15.0] - 2026-01-28
+
+- No significant changes
+
 ## [0.14.0] - 2026-01-15
 
 ### Added